Just accomplished to get the CSimulation running outputting the right results AND the synthesis working without an error. Really great !!!! Just don't know whether it works, but thats secondary :-)

+++ src/fpgaPlayground/HLS/AFAalpha/src/AFARandom.c
@@ -19,13 +19,14 @@
 //!        should give 2^19937-1 random numbers per seed
 
 #include "AFARandom.h"
+#include "AFATypes.h"
 
 #include <assert.h>
 
 
 
 
-unsigned long m_mt[RANDOM_N]; // the array for the state vector 
+uint32_t m_mt[ RANDOM_N ]; // the array for the state vector 
 int m_mti = RANDOM_N;
